Implements verify consensus fault syscall and reorg vm crates (#386)
* added default runtime to interpreter crate

* Fixed transfer test and gas syscalls

* Added std error to syscalls trait and made requested changes

* Made downcasting change

* made requested changes by returing if none before expensive checks
