+++ vm/interpreter/tests/transfer_test.rs
@@ -5,11 +5,10 @@ use actor::{init, ACCOUNT_ACTOR_CODE_ID, INIT_ACTOR_ADDR};
 use address::Address;
 use cid::multihash::{Blake2b256, Identity};
 use db::MemoryDB;
-use default_runtime::{internal_send, DefaultRuntime};
+use interpreter::{internal_send, DefaultRuntime, DefaultSyscalls};
 use ipld_blockstore::BlockStore;
 use ipld_hamt::Hamt;
 use message::UnsignedMessage;
-use runtime::DefaultSyscalls;
 use state_tree::StateTree;
 use vm::{ActorState, Serialized};
 
@@ -92,10 +91,12 @@ fn transfer_test() {
         .build()
         .unwrap();
 
+    let default_syscalls = DefaultSyscalls::new(&store);
+
     let mut runtime = DefaultRuntime::new(
         &mut state,
         &store,
-        DefaultSyscalls,
+        &default_syscalls,
         0,
         &message,
         0,
